refactor: move endpoint services to own directory
