+++ api/server/services/Endpoints/EndpointService.js
@@ -1,4 +1,4 @@
-const { EModelEndpoint } = require('../routes/endpoints/schemas');
+const { EModelEndpoint } = require('~/server/routes/endpoints/schemas');
 
 const {
   OPENAI_API_KEY: openAIApiKey,
