+++ api/server/services/ModelService.js
@@ -1,11 +1,11 @@
 const HttpsProxyAgent = require('https-proxy-agent');
 const axios = require('axios');
 const Keyv = require('keyv');
-const { isEnabled } = require('../utils');
-const { extractBaseURL } = require('../../utils');
-const keyvRedis = require('../../cache/keyvRedis');
-// const { getAzureCredentials, genAzureChatCompletion } = require('../../utils/');
-const { openAIApiKey, userProvidedOpenAI } = require('./EndpointService').config;
+const { isEnabled } = require('~/server/utils');
+const { extractBaseURL } = require('~/utils');
+const keyvRedis = require('~/cache/keyvRedis');
+// const { getAzureCredentials, genAzureChatCompletion } = require('~/utils/');
+const { openAIApiKey, userProvidedOpenAI } = require('./Endpoints/EndpointService').config;
 
 const modelsCache = isEnabled(process.env.USE_REDIS)
   ? new Keyv({ store: keyvRedis })
